Pull out make_grid into new file utils.py
